Ahora tambien calcula la edad del universo siempre y cuando el usuario haya metido los datos de Mundo disco primero y criatura.

+++ Lab6_KennethNuñez/src/lab6_kennethnuñez/Universo.java
@@ -3,7 +3,7 @@ package lab6_kennethnuñez;
 public class Universo {
     private String UniversoCode;
     private int UniverseAge;
-    private int peso;
+    private double peso;
 
     public Universo() {
     }
@@ -30,7 +30,7 @@ public class Universo {
         this.UniverseAge = UniverseAge;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
